requirements.txt
###### Requirements without Version Specifiers ######
django~=4.0.3
djangorestframework~=3.13.1
django-heroku
django-cors-headers
pytest~=7.1.2
pipe~=1.6.0
mysqlclient